src/third_party/blink/web_tests/external/wpt/fetch/sec-metadata/redirect/multiple-redirect-cross-site.tentative.https.sub-expected.txt

This is a testharness.js-based test.
FAIL Same-Origin -> Cross-Site -> Same-Origin redirect assert_equals: site expected "cross-site" but got "same-origin"
Harness: the test ran to completion.
